requirements.txt
streamlit
google-generativeai
pillow
pdf2image
python-dotenv